Feat: Update version to 4.11.17-stable, refactor code structure, and enhance CLI functionality
