+++ README.md
@@ -31,10 +31,10 @@ ContainDB is a CLI tool written in Go that automates the installation and manage
 Option 1: Download and install via Debian package  
 ```bash
 # download latest .deb from Releases
-wget https://github.com/AnkanSaha/ContainDB/releases/download/v3.11.18-stable/containDB_3.11.18-stable_amd64.deb
+wget https://github.com/AnkanSaha/ContainDB/releases/download/v4.11.17-stable/containDB_4.11.17-stable_amd64.deb
 
 # install the package
-sudo dpkg -i containDB_3.11.18-stable_amd64.deb
+sudo dpkg -i containDB_4.11.17-stable_amd64.deb
 ```
 
 Option 2: Build from source  
@@ -54,7 +54,7 @@ sudo mv ./bin/containDB /usr/local/bin/
 
 ## Quick Start
 
-Run the CLI with root privileges:
+Run the CLI with Root privileges:
 ```bash
 sudo containDB
 ```
