+++ Scripts/BinBuilder.sh
@@ -8,7 +8,7 @@ BINARY_PATH="./bin/ContainDB"
 BUILD_OUTPUT_DIR=$(dirname "$BINARY_PATH")
 
 # Check if Go is installed
-if ! command -v go &> /dev/null; then
+if ! command -v go &>/dev/null; then
   echo "Go not found. Installing Go via snap..."
   sudo snap install go --classic
 else
@@ -20,6 +20,6 @@ mkdir -p "$BUILD_OUTPUT_DIR"
 
 # Build the Go project with -o flag
 echo "Building the project..."
-go build -o "$BINARY_PATH" ./src/Root
+go build -o "$BINARY_PATH" ./src/Core
 
 echo "Build complete. Binary available at $BINARY_PATH"
