+++ src/base/BaseCaseHandler.go
@@ -1,9 +1,7 @@
-package main
+package base;
 
 import (
 	"fmt"
-	"os"
-	"os/signal"
 
 	"ContainDB/src/Docker"
 	"ContainDB/src/tools"
@@ -11,61 +9,7 @@ import (
 	"github.com/manifoldco/promptui"
 )
 
-func main() {
-	VERSION := "3.11.16-stable"
-
-	// handle version flag without requiring sudo
-	if len(os.Args) > 1 && os.Args[1] == "--version" {
-		fmt.Println("ContainDB CLI Version:", VERSION)
-		return
-	}
-
-	// Replace Ctrl+C handler to avoid triggering on normal exit
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, os.Interrupt)
-	go func() {
-		<-sigCh
-		fmt.Println("\n⚠️ Interrupt received, rolling back...")
-		tools.Cleanup()
-		os.Exit(1)
-	}()
-
-	// require sudo
-	if os.Geteuid() != 0 {
-		fmt.Println("❌ Please run this program with sudo")
-		os.Exit(1)
-	}
-
-	if !Docker.IsDockerInstalled() {
-		fmt.Println("❌ Docker is not installed. Without Docker the tool cannot run.")
-		installPrompt := promptui.Select{
-			Label: "Would you like to install Docker now?",
-			Items: []string{"Yes", "No", "Exit"},
-		}
-		_, choice, err := installPrompt.Run()
-		if err != nil || choice != "Yes" {
-			fmt.Println("Exiting. Please install Docker manually and rerun.")
-			os.Exit(1)
-		}
-		fmt.Println(("Checking system requirements..."))
-		Docker.CheckSystemRequirements() // Check system requirements before installing Docker
-		err = Docker.InstallDocker()
-		if err != nil {
-			fmt.Println("Failed to install Docker:", err)
-			return
-		}
-		fmt.Println("Docker installed successfully! Please restart the terminal or log out & log in again.")
-		return
-	}
-
-	err := Docker.CreateDockerNetworkIfNotExists()
-	if err != nil {
-		fmt.Println("Failed to create Docker network:", err)
-		return
-	}
-
-	// Show welcome banner
-	ShowBanner()
+func BaseCaseHandler() {
 
 	// Top-level action menu
 	actionPrompt := promptui.Select{
@@ -135,4 +79,5 @@ func main() {
 		fmt.Println("Goodbye!")
 		return
 	}
+
 }
