+++ src/base/DatabaseSelector.go
@@ -1,4 +1,4 @@
-package main
+package base;
 
 import (
 	"ContainDB/src/tools"
