+++ src/base/flagHandler.go
@@ -1,4 +1,4 @@
-package main
+package base;
 
 import (
 	"ContainDB/src/Docker"
@@ -7,8 +7,8 @@ import (
 )
 
 // flagHandler handles command line flags for the ContainDB CLI
-func flagHandler() {
-	VERSION := "3.11.18-stable"
+func FlagHandler() {
+	VERSION := "4.11.17-stable"
 	// handle version flag without requiring sudo
 	if len(os.Args) > 1 && os.Args[1] == "--version" {
 		fmt.Println("ContainDB CLI Version:", VERSION)
